pyproject.toml
[project]
name = "14mv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "opencv-python-headless>=4.11.0.86",
    "pillow>=11.1.0",
    "pygetwindow>=0.0.9",
    "pyqt5-tools>=5.15.9.3.3",
    "pytest>=8.3.5",
]
